Move partition col names & data schema methods to Metadata
