+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/actions/Metadata.java
@@ -15,9 +15,8 @@
  */
 package io.delta.kernel.internal.actions;
 
-import java.util.Collections;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
+import java.util.stream.Collectors;
 import static java.util.Objects.requireNonNull;
 
 import io.delta.kernel.client.TableClient;
@@ -29,6 +28,7 @@ import io.delta.kernel.types.*;
 import io.delta.kernel.internal.lang.Lazy;
 import io.delta.kernel.internal.util.VectorUtils;
 import static io.delta.kernel.internal.util.InternalUtils.requireNonNull;
+import static io.delta.kernel.internal.util.Preconditions.checkArgument;
 
 public class Metadata {
 
@@ -82,6 +82,10 @@ public class Metadata {
     private final Optional<Long> createdTime;
     private final MapValue configurationMapValue;
     private final Lazy<Map<String, String>> configuration;
+    // Partition column names in lower case.
+    private final Lazy<Set<String>> partitionColNames;
+    // Logical data schema excluding partition columns
+    private final Lazy<StructType> dataSchema;
 
     public Metadata(
         String id,
@@ -103,6 +107,11 @@ public class Metadata {
         this.createdTime = createdTime;
         this.configurationMapValue = requireNonNull(configurationMapValue, "configuration is null");
         this.configuration = new Lazy<>(() -> VectorUtils.toJavaMap(configurationMapValue));
+        this.partitionColNames = new Lazy<>(() -> loadPartitionColNames());
+        this.dataSchema = new Lazy<>(() ->
+            new StructType(schema.fields().stream()
+                .filter(field -> !partitionColNames.get().contains(field.getName().toLowerCase(Locale.ROOT)))
+                .collect(Collectors.toList())));
     }
 
     public String getSchemaString() {
@@ -115,6 +124,16 @@ public class Metadata {
 
     public ArrayValue getPartitionColumns() {
         return partitionColumns;
+    }
+
+    /** Set of lowercase partition column names */
+    public Set<String> getPartitionColNames() {
+        return partitionColNames.get();
+    }
+
+    /** The logical data schema which excludes partition columns */
+    public StructType getDataSchema() {
+        return dataSchema.get();
     }
 
     public String getId() {
@@ -143,5 +162,22 @@ public class Metadata {
 
     public Map<String, String> getConfiguration() {
         return Collections.unmodifiableMap(configuration.get());
+    }
+
+    /**
+     * Helper method to load the partition column names.
+     */
+    private Set<String> loadPartitionColNames() {
+        ColumnVector partitionColNameVector = partitionColumns.getElements();
+        Set<String> partitionColumnNames = new HashSet<>();
+        for (int i = 0; i < partitionColumns.getSize(); i++) {
+            checkArgument(!partitionColNameVector.isNullAt(i),
+                "Expected a non-null partition column name");
+            String partitionColName = partitionColNameVector.getString(i);
+            checkArgument(partitionColName != null && !partitionColName.isEmpty(),
+                "Expected non-null and non-empty partition column name");
+            partitionColumnNames.add(partitionColName.toLowerCase(Locale.ROOT));
+        }
+        return Collections.unmodifiableSet(partitionColumnNames);
     }
 }
